Add support for several more CPUs
The list of supported signature types is now complete and fully updated, according to the tables given in 'https://software.intel.com/sites/default/files/managed/22/0d/335592-sdm-vol-4.pdf'.
As to whether the given MSRs are enough for the first Atom-Generations remains to be seen (i.e., Silvermont, Merrified, Moorefield and Airmont). For now I included those CPU types mainly for completeness. I will check that in a future update.