Undo a previous commit due to the changes made being redundant by now
The removed code checked the msr_table (whose values were previously inserted manually into the table), and afterwards checked the value again by retrieving it from the hardware on the working machine.
This step is now redundant, as the values of the msr_table are no longer inserted by hand, but instead the program simply retrieves from the processor as to whether a specific RAPL domain is in fact supported.