Make time interval of energy probes dependant on max. specified power in the corresp. msr
The maximum power is derived from the electrical specification of the package domain (i.e. 'MSR_PKG_POWER_INFO' -- see Intel Architectures Software Developer's Manual, Volume 3B: System Programming Guide, Part 2, Section 14.9.3, Fig. 14-34).
The concrete value is the maximum value between 'Thermal Spec Power' and 'Maximum Power'. If both of these are equal to zero, an arbitrary (high) value of 200 Watt is taken. For example, on my current working machine, 15W were retrieved as max. power consumption from the msr. This translates to nearly 5 hours, until an overflow comes to pass. With this commit, half of the time is now used as interval value for making probes in each domain of the total consumed energy.
The commit fixes the issue as described in #12 on the GitHub repo.