Drop root privileges after opening the fd(s)
This still needs some modifications, because as of now, the functioning only works when the program is invoked without root-privileges (i.e., running the program without sudo). This is due to getuid() and geteuid() returning 0 when running the program as root - the real user ID is that of the user who has invoked the program (i.e., root in this case). This can be easily solved by setting an arbitrary non-zero value, though, I'll leave it as it is for now.