Update 'make dist' to retain the dir structure and to include the testfiles
The following three steps are now required in order to create a new deb-file:
1.) 'make dist'
2.) 'dh_make -f cpu-energy-meter-1.1-dev.tar.gz -p cpu-energy-meter-1.1',
    with the version number changed accordingly
3.) dpkg-buildpackage -us -uc

Note that for the third command to be executed successfully, the debian/changelog
file needs to be changed such that the version number corresponds with the version
number given in the -p argument in the second line, and the distribution is set to
something that is accepted by dh_make like e.g. unstable.

(PS: I'm currently writing a more comprehensive short guide on how to set up a
Debian Package file, which could eventually be shared some time later)