Copy ruby-scriptfile from the CMock submodule into the main-project
The ruby-script 'create_makefile.rb' in the CMock-project was apparently written and
tested only for the example-projects in 'CMock/examples', because in line 3 the
'CMock_DIR'-variable has the flaw that it only works for projects with such a directory-
structure:
'CMOCK_DIR = File.expand_path(ENV.fetch('CMOCK_DIR', File.join(ABS_ROOT, '..', '..')))'.
Unfortunately, this path-statement doesn't work for the 'cpu-energy-meter'-project, in
which CMock is only used as a submodule. So in order to avoid making changes to a file in
an own submodule, the (temporary) solution was to copy the according file into the main-
project where the appropriate changes can be easily made.