Use REUSE-style license and copyright declarations for source files
Contrary to the main license file of the repo,
the declared license of each of these file is actually BSD-3-Clause
instead of BSD-2-Clause (detected with scancode and manually confirmed).
This was already the case in the original Intel code.

So we also have to update the license declaration in debian/copyright.