Relicense two files from BSD-2-Clause to BSD-3-Clause
This is possible because BSD-3-Clause fulfills all the requirements of
BSD-2-Clause: We only need to include the copyright notice, the
disclaimer (identical in both license variants) and the list of
conditions from the license (and the list of conditions in BSD-3-Clause
is a superset of the list of conditions in BSD-2-Clause).
There is no requirement in BSD-2-Clause to retain the exact license
text, and no prevention of adding additional conditions.

BSD-3-Clause probably was the originally intended license anyway
given that all source files explicitly had BSD-3-Clause
and only the main LICENSE file of the project was BSD-2-Clause.